+++ beets/autotag/__init__.py
@@ -33,13 +33,14 @@ log = logging.getLogger('beets')
 
 # Additional utilities for the main interface.
 
-def albums_in_dir(path):
+def albums_in_dir(path, ignore=()):
     """Recursively searches the given directory and returns an iterable
     of (path, items) where path is a containing directory and items is
     a list of Items that is probably an album. Specifically, any folder
-    containing any media files is an album.
+    containing any media files is an album. Directories and file names
+    that match the glob patterns in ``ignore`` are skipped.
     """
-    for root, dirs, files in sorted_walk(path):
+    for root, dirs, files in sorted_walk(path, ignore):
         # Get a list of items in the directory.
         items = []
         for filename in files: