+++ test/test_ui.py
@@ -431,7 +431,7 @@ class ImportTest(unittest.TestCase):
         self.assertRaises(ui.UserError, commands.import_files,
                           None, [], False, False, False, None, False, False,
                           False, False, True, False, None, False, True, None,
-                          False)
+                          False, [])
 
 class InputTest(unittest.TestCase):
     def setUp(self):